chore: rewrite back to .js
